refactor(protocol): split demo.ts into per-area scenario modules
Design choice: kept a simple orchestrator plus per-area scenario functions.

This avoids adding a registry abstraction for a mechanical split.

The refactor preserves the existing scenario order and 144-pass demo summary.

Shared assertions and the receipt fixture moved into dedicated helpers.

The demo did not need new scenarios because this is decomposition only.

check-invariants now scans scripts/demo/ for index-only imports and coverage.
