+++ packages/protocol/scripts/demo.ts
@@ -12,1645 +12,24 @@
 // laid out as a narrative so a reviewer doesn't have to read fast-check
 // arbitraries to be convinced the moat fires on the right inputs.
 
-import { inspect } from "node:util";
-
-import {
-  type ApprovalRequest,
-  type AuditEventRecord,
-  apiBootstrapFromJson,
-  apiBootstrapToJson,
-  approvalDecisionRequestFromJson,
-  approvalDecisionRequestToJsonValue,
-  approvalDecisionResponseFromJson,
-  approvalDecisionResponseToJsonValue,
-  approvalRequestCreateRequestFromJson,
-  approvalRequestCreateRequestToJsonValue,
-  approvalRequestCreateResponseFromJson,
-  approvalRequestCreateResponseToJsonValue,
-  approvalRequestFromJsonValue,
-  approvalRequestToJson,
-  approvalRequestToJsonValue,
-  approvalSubmitRequestFromJson,
-  asAgentId,
-  asAgentSlug,
-  asApiToken,
-  asApprovalClaimId,
-  asApprovalId,
-  asApprovalRequestId,
-  asApprovalRole,
-  asApprovalTokenId,
-  asBudgetId,
-  asCredentialHandleId,
-  asCredentialScope,
-  asIdempotencyKey,
-  asMerkleRootHex,
-  asMicroUsd,
-  asProviderKind,
-  asReceiptId,
-  asSignerIdentity,
-  asTaskId,
-  asThreadId,
-  asThreadSpecRevisionId,
-  asTimestampMs,
-  asToolCallId,
-  asWriteId,
-  BUDGET_SCOPE_VALUES,
-  type BudgetSetAuditPayload,
-  type BudgetThresholdCrossedAuditPayload,
-  type CostEventAuditPayload,
-  canonicalJSON,
-  computeAuditEventHash,
-  costAuditPayloadFromJsonValue,
-  costAuditPayloadToBytes,
-  costAuditPayloadToJsonValue,
-  createCredentialHandle,
-  credentialHandleFromJson,
-  credentialHandleToJson,
-  type EventLsn,
-  FrozenArgs,
-  GENESIS_PREV_HASH,
-  INITIAL_VERIFIER_STATE,
-  isAllowedLoopbackHost,
-  isApprovalRole,
-  isBudgetId,
-  isBudgetScope,
-  isCostAuditEventKind,
-  isLoopbackRemoteAddress,
-  isMicroUsd,
-  isRunnerFailureCode,
-  isStreamEventKind,
-  isWsFrameType,
-  lsnFromV1Number,
-  MAX_AUDIT_CHAIN_BATCH_SIZE,
-  MAX_BUDGET_THRESHOLD_BPS,
-  MAX_BUDGET_THRESHOLDS,
-  MAX_COST_EVENT_AMOUNT_MICRO_USD,
-  MAX_COST_MODEL_BYTES,
-  MAX_TOOL_CALLS_PER_RECEIPT,
-  MAX_WEBAUTHN_ASSERTION_FIELD_BYTES,
-  MINIMUM_PROTOCOL_VERSION_FOR_PROVIDER_KIND,
-  type ReceiptSnapshot,
-  RUNNER_FAILURE_CODE_VALUES,
-  receiptFromJson,
-  receiptToJson,
-  routeErrorFromJson,
-  routeErrorToJsonValue,
-  runnerEventFromJson,
-  runnerEventToJsonValue,
-  runnerSpawnRequestFromJson,
-  runnerSpawnRequestToJsonValue,
-  SanitizedString,
-  STREAM_EVENT_KIND_VALUES,
-  serializeAuditEventRecordForHash,
-  sha256Hex,
-  signedApprovalTokenToJsonValue,
-  type Thread,
-  type ThreadSpecEditedAuditPayload,
-  type ThreadStreamEvent,
-  threadAuditPayloadToBytes,
-  threadCreateRequestFromJson,
-  threadCreateRequestToJsonValue,
-  threadFromJson,
-  threadGetResponseFromJson,
-  threadGetResponseToJsonValue,
-  threadListResponseFromJson,
-  threadListResponseToJsonValue,
-  threadMutationResponseFromJson,
-  threadMutationResponseToJsonValue,
-  threadSpecContentHash,
-  threadSpecEditRequestFromJson,
-  threadSpecEditRequestToJsonValue,
-  threadStatusChangeRequestFromJson,
-  threadStatusChangeRequestToJsonValue,
-  threadToJson,
-  validateApprovalSubmitRequest,
-  validateBudgetSetAuditPayload,
-  validateBudgetThresholdCrossedAuditPayload,
-  validateCostAuditPayloadForKind,
-  validateCostEventAuditPayload,
-  validateMerkleRootRecord,
-  validateReceiptBudget,
-  validateThreadSpecEditedAuditPayload,
-  validateThreadSpecRevisionChain,
-  validateThreadStatusFold,
-  validateThreadStreamEvent,
-  verifyChain,
-  verifyChainIncremental,
-  WS_FRAME_TYPE_VALUES,
-} from "../src/index.ts";
-
-const ANSI = {
-  reset: "\x1b[0m",
-  dim: "\x1b[2m",
-  bold: "\x1b[1m",
-  green: "\x1b[32m",
-  red: "\x1b[31m",
-  yellow: "\x1b[33m",
-  cyan: "\x1b[36m",
-};
-
-let passed = 0;
-let failed = 0;
-const textDecoder = new TextDecoder();
-
-function header(num: number, title: string): void {
-  console.log("");
-  console.log(`${ANSI.bold}${ANSI.cyan}── Scenario ${num}: ${title}${ANSI.reset}`);
-}
-
-function expectThrows(fn: () => unknown, expectedFragment: RegExp): void {
-  try {
-    const result = fn();
-    console.log(`  ${ANSI.red}FAIL${ANSI.reset} expected throw, got: ${JSON.stringify(result)}`);
-    failed++;
-  } catch (err) {
-    const msg = err instanceof Error ? err.message : String(err);
-    if (expectedFragment.test(msg)) {
-      console.log(
-        `  ${ANSI.green}PASS${ANSI.reset} threw: ${ANSI.dim}${msg.slice(0, 100)}${ANSI.reset}`,
-      );
-      passed++;
-    } else {
-      console.log(
-        `  ${ANSI.red}FAIL${ANSI.reset} threw wrong message: ${msg}\n` +
-          `       expected to match: ${expectedFragment}`,
-      );
-      failed++;
-    }
-  }
-}
-
-function expectEqual<T>(label: string, actual: T, expected: T): void {
-  const ok = JSON.stringify(actual) === JSON.stringify(expected);
-  if (ok) {
-    console.log(
-      `  ${ANSI.green}PASS${ANSI.reset} ${label} = ${ANSI.dim}${JSON.stringify(actual)}${ANSI.reset}`,
-    );
-    passed++;
-  } else {
-    console.log(
-      `  ${ANSI.red}FAIL${ANSI.reset} ${label}\n` +
-        `       expected: ${JSON.stringify(expected)}\n` +
-        `       actual:   ${JSON.stringify(actual)}`,
-    );
-    failed++;
-  }
-}
-
-function expectCodecRoundTrip<T>(
-  label: string,
-  input: unknown,
-  fromJson: (value: unknown) => T,
-  toJsonValue: (value: T) => unknown,
-): void {
-  const serialized = toJsonValue(fromJson(input));
-  const reparsed = fromJson(JSON.parse(canonicalJSON(serialized)));
-  expectEqual(label, canonicalJSON(toJsonValue(reparsed)), canonicalJSON(serialized));
-}
-
-function nonNull<T>(value: T | null | undefined, label: string): T {
-  if (value === null || value === undefined) {
-    throw new Error(`demo fixture missing required value: ${label}`);
-  }
-  return value;
-}
-
-function expectChainResult(
-  label: string,
-  actual: ReturnType<typeof verifyChain>,
-  expectedCode: string | "ok",
-): void {
-  const actualCode = actual.ok ? "ok" : actual.code;
-  if (actualCode === expectedCode) {
-    console.log(
-      `  ${ANSI.green}PASS${ANSI.reset} ${label} = ${ANSI.dim}${actualCode}${ANSI.reset}`,
-    );
-    passed++;
-  } else {
-    console.log(
-      `  ${ANSI.red}FAIL${ANSI.reset} ${label} expected ${expectedCode}, got ${actualCode}`,
-    );
-    failed++;
-  }
-}
-
-function expectIncrementalResult(
-  label: string,
-  actual: ReturnType<typeof verifyChainIncremental>,
-  expectedCode: string | "ok",
-): void {
-  const actualCode = actual.ok ? "ok" : actual.code;
-  if (actualCode === expectedCode) {
-    console.log(
-      `  ${ANSI.green}PASS${ANSI.reset} ${label} = ${ANSI.dim}${actualCode}${ANSI.reset}`,
-    );
-    passed++;
-  } else {
-    console.log(
-      `  ${ANSI.red}FAIL${ANSI.reset} ${label} expected ${expectedCode}, got ${actualCode}`,
-    );
-    failed++;
-  }
-}
-
-function expectBudgetFailure(
-  label: string,
-  actual: ReturnType<typeof validateReceiptBudget>,
-  expectedFragment: RegExp,
-): void {
-  if (!actual.ok && expectedFragment.test(actual.reason)) {
-    console.log(
-      `  ${ANSI.green}PASS${ANSI.reset} ${label} = ${ANSI.dim}${actual.reason}${ANSI.reset}`,
-    );
-    passed++;
-  } else {
-    console.log(
-      `  ${ANSI.red}FAIL${ANSI.reset} ${label} expected reason matching ${expectedFragment}`,
-    );
-    failed++;
-  }
-}
+import { runApprovalRouteScenarios } from "./demo/approval-route-scenarios.ts";
+import { runCoreScenarios } from "./demo/core-scenarios.ts";
+import { runCostScenarios } from "./demo/cost-scenarios.ts";
+import { runCredentialRunnerScenarios } from "./demo/credential-runner-scenarios.ts";
+import { ANSI, printSummaryAndExit } from "./demo/harness.ts";
+import { runIpcScenarios } from "./demo/ipc-scenarios.ts";
+import { touchPublicSurfaceSentinels } from "./demo/public-surface-sentinels.ts";
+import { runThreadScenarios } from "./demo/thread-scenarios.ts";
 
 console.log(`${ANSI.bold}@wuphf/protocol — moat demo${ANSI.reset}`);
 console.log(`${ANSI.dim}Each scenario: input → expected behavior → actual.${ANSI.reset}`);
 
-// ──────────────────────────────────────────────────────────────────────────
-header(1, "FrozenArgs is content-addressed, key-order-independent");
-// ──────────────────────────────────────────────────────────────────────────
-const argsA = FrozenArgs.freeze({ b: 2, a: 1 });
-const argsB = FrozenArgs.freeze({ a: 1, b: 2 });
-expectEqual("hash equality across key order", argsA.hash === argsB.hash, true);
-expectEqual("canonical JSON matches", argsA.canonicalJson, '{"a":1,"b":2}');
+runCoreScenarios();
+runIpcScenarios();
+runThreadScenarios();
+runCostScenarios();
+runCredentialRunnerScenarios();
+runApprovalRouteScenarios();
+touchPublicSurfaceSentinels();
 
-// ──────────────────────────────────────────────────────────────────────────
-header(2, "SanitizedString strips weaponized invisible code points");
-// ──────────────────────────────────────────────────────────────────────────
-expectEqual(
-  "U+180E (Mongolian Vowel Separator) stripped",
-  SanitizedString.fromUnknown("ad᠎min").value,
-  "admin",
-);
-expectEqual("U+2060 (Word Joiner) stripped", SanitizedString.fromUnknown("ev⁠il").value, "evil");
-expectEqual(
-  "U+202E (Right-to-Left Override) stripped",
-  SanitizedString.fromUnknown("evil‮txt.exe").value,
-  "eviltxt.exe",
-);
-expectEqual(
-  "U+FB01 (ﬁ ligature) NFKC-normalizes to 'fi'",
-  SanitizedString.fromUnknown("ﬁle").value,
-  "file",
-);
-
-// ──────────────────────────────────────────────────────────────────────────
-header(3, "SanitizedString rejects untrusted graph BEFORE side-effects fire");
-// ──────────────────────────────────────────────────────────────────────────
-let getterFired = false;
-const adversarialInput = {
-  get tricky() {
-    getterFired = true;
-    return "should-never-be-read";
-  },
-};
-expectThrows(
-  () => SanitizedString.fromUnknown(adversarialInput),
-  /accessor|getter|toJSON|non-plain/i,
-);
-expectEqual("getter side-effect did NOT fire", getterFired, false);
-expectThrows(
-  () => SanitizedString.fromUnknown({ toJSON: () => "spoofed" }),
-  /toJSON|accessor|non-plain/i,
-);
-expectThrows(
-  () => SanitizedString.fromUnknown(new Map([["k", "v"]])),
-  /Map|non-plain|class instance/i,
-);
-
-// ──────────────────────────────────────────────────────────────────────────
-header(4, "Canonical JSON rejects prototype pollution attempts");
-// ──────────────────────────────────────────────────────────────────────────
-expectThrows(
-  () => SanitizedString.fromUnknown(JSON.parse('{"__proto__":{"polluted":true},"ok":1}')),
-  /__proto__|forbidden|prototype/,
-);
-expectThrows(
-  () => canonicalJSON(JSON.parse('{"__proto__":{"polluted":true},"ok":1}')),
-  /canonicalJSON: forbidden key "__proto__"/,
-);
-expectThrows(
-  () => SanitizedString.fromUnknown(JSON.parse('{"constructor":{"polluted":true}}')),
-  /constructor|forbidden|prototype/,
-);
-expectThrows(
-  () => FrozenArgs.freeze(JSON.parse('{"prototype":{"polluted":true}}')),
-  /canonicalJSON: forbidden key "prototype"/,
-);
-
-// ──────────────────────────────────────────────────────────────────────────
-header(5, "Audit chain catches tampering with a typed failure code");
-// ──────────────────────────────────────────────────────────────────────────
-function buildChain(n: number): AuditEventRecord[] {
-  const out: AuditEventRecord[] = [];
-  let prev = GENESIS_PREV_HASH;
-  for (let i = 0; i < n; i++) {
-    const partial: AuditEventRecord = {
-      seqNo: lsnFromV1Number(i),
-      timestamp: new Date(2026, 4, 8, 0, 0, i),
-      prevHash: prev,
-      eventHash: GENESIS_PREV_HASH,
-      payload: { kind: "receipt_created", body: new TextEncoder().encode(`event-${i}`) },
-    };
-    const eventHash = computeAuditEventHash(partial);
-    out.push({ ...partial, eventHash });
-    prev = eventHash;
-  }
-  return out;
-}
-
-expectChainResult("clean 5-record chain", verifyChain(buildChain(5)), "ok");
-
-const tamperedChain = buildChain(5);
-const recordToTamper = nonNull(tamperedChain[2], "tamperedChain[2]");
-tamperedChain[2] = {
-  ...recordToTamper,
-  payload: { kind: "receipt_finalized", body: new TextEncoder().encode("malicious") },
-};
-expectChainResult(
-  "chain with tampered payload at seq 2",
-  verifyChain(tamperedChain),
-  "event_hash_mismatch",
-);
-
-const unknownKindRecord = nonNull(buildChain(1)[0], "unknownKindRecord");
-expectChainResult(
-  "chain with unknown payload kind",
-  verifyChain([
-    {
-      ...unknownKindRecord,
-      payload: {
-        ...unknownKindRecord.payload,
-        kind: "made_up_kind" as AuditEventRecord["payload"]["kind"],
-      },
-    },
-  ]),
-  "serialization_threw",
-);
-
-const oversizedBodyRecord = nonNull(buildChain(1)[0], "oversizedBodyRecord");
-expectChainResult(
-  "chain with oversized audit event body",
-  verifyChain([
-    {
-      ...oversizedBodyRecord,
-      payload: {
-        ...oversizedBodyRecord.payload,
-        body: new Uint8Array(1 * 1024 * 1024 + 1),
-      },
-    },
-  ]),
-  "serialization_threw",
-);
-
-const sparseChain = buildChain(3);
-Reflect.deleteProperty(sparseChain, "1");
-expectChainResult("chain with deleted record at seq 1", verifyChain(sparseChain), "missing_record");
-
-// ──────────────────────────────────────────────────────────────────────────
-header(6, "Audit chain verifies incrementally without retaining the full log");
-// ──────────────────────────────────────────────────────────────────────────
-const incrementalChain = buildChain(12);
-const firstIncrementalBatch = verifyChainIncremental(
-  INITIAL_VERIFIER_STATE,
-  incrementalChain.slice(0, 5),
-);
-expectIncrementalResult("first 5-record batch", firstIncrementalBatch, "ok");
-if (firstIncrementalBatch.ok) {
-  const secondIncrementalBatch = verifyChainIncremental(
-    firstIncrementalBatch.state,
-    incrementalChain.slice(5),
-  );
-  expectIncrementalResult("remaining 7-record batch", secondIncrementalBatch, "ok");
-  if (secondIncrementalBatch.ok) {
-    const allAtOnce = verifyChain(incrementalChain);
-    expectEqual(
-      "incremental last hash matches verifyChain",
-      allAtOnce.ok && !allAtOnce.empty
-        ? secondIncrementalBatch.state.expectedPrev === allAtOnce.lastEventHash
-        : false,
-      true,
-    );
-  }
-}
-expectIncrementalResult(
-  "oversized verifier batch",
-  verifyChainIncremental(
-    INITIAL_VERIFIER_STATE,
-    new Array<AuditEventRecord>(MAX_AUDIT_CHAIN_BATCH_SIZE + 1),
-  ),
-  "batch_too_large",
-);
-
-// ──────────────────────────────────────────────────────────────────────────
-header(7, "Receipt budgets reject runaway task fanout");
-// ──────────────────────────────────────────────────────────────────────────
-const budgetFixture = buildValidReceipt();
-const runawayToolCall = nonNull(budgetFixture.toolCalls[0], "budgetFixture.toolCalls[0]");
-expectEqual("normal receipt budget", validateReceiptBudget(budgetFixture), { ok: true });
-expectBudgetFailure(
-  "toolCalls length cap",
-  validateReceiptBudget({
-    ...budgetFixture,
-    toolCalls: Array.from({ length: MAX_TOOL_CALLS_PER_RECEIPT + 1 }, () => runawayToolCall),
-  }),
-  /toolCalls.*exceeds budget/,
-);
-
-// ──────────────────────────────────────────────────────────────────────────
-header(8, "FrozenArgs JSON envelope rejects smuggled siblings");
-// ──────────────────────────────────────────────────────────────────────────
-const validReceiptJson = receiptToJson(buildValidReceipt());
-const parsed = JSON.parse(validReceiptJson) as {
-  toolCalls: Array<{ inputs: Record<string, unknown> }>;
-};
-const firstToolCall = nonNull(parsed.toolCalls[0], "parsed.toolCalls[0]");
-firstToolCall.inputs = { ...firstToolCall.inputs, evilShadow: "smuggled" };
-expectThrows(() => receiptFromJson(JSON.stringify(parsed)), /evilShadow.*not allowed/);
-
-// ──────────────────────────────────────────────────────────────────────────
-header(9, "ExternalWrite per-state invariants");
-// ──────────────────────────────────────────────────────────────────────────
-const tamperedApplied = JSON.parse(receiptToJson(buildValidReceipt())) as {
-  writes: Array<Record<string, unknown>>;
-};
-const firstWrite = nonNull(tamperedApplied.writes[0], "tamperedApplied.writes[0]");
-firstWrite.appliedDiff = null;
-expectThrows(
-  () => receiptFromJson(JSON.stringify(tamperedApplied)),
-  /appliedDiff.*null is invalid for state "applied"/,
-);
-
-// ──────────────────────────────────────────────────────────────────────────
-header(10, "Approval token writeId binding catches cross-write authorization");
-// ──────────────────────────────────────────────────────────────────────────
-const writeIdMismatch = JSON.parse(receiptToJson(buildValidReceipt())) as {
-  writes: Array<Record<string, unknown>>;
-};
-const writeWithToken = nonNull(writeIdMismatch.writes[0], "writeIdMismatch.writes[0]");
-const tokenWithWrongWriteId = JSON.parse(JSON.stringify(writeWithToken.approvalToken)) as {
-  claim: Record<string, unknown>;
-};
-tokenWithWrongWriteId.claim.writeId = "write_wrong_target";
-writeWithToken.approvalToken = tokenWithWrongWriteId;
-expectThrows(() => receiptFromJson(JSON.stringify(writeIdMismatch)), /writeId.*must match/);
-
-// ──────────────────────────────────────────────────────────────────────────
-header(11, "ApprovalSubmitRequest cross-field validator (IPC layer)");
-// ──────────────────────────────────────────────────────────────────────────
-const validReceipt = buildValidReceipt();
-const validToken = nonNull(validReceipt.approvals[0], "validReceipt.approvals[0]").signedToken;
-const goodReq = {
-  receiptId: validReceipt.id,
-  approvalToken: validToken,
-  idempotencyKey: asIdempotencyKey("submit-01"),
-};
-expectEqual("matched receiptId/claim", validateApprovalSubmitRequest(goodReq), { ok: true });
-
-const badReq = {
-  ...goodReq,
-  receiptId: asReceiptId("01ARZ3NDEKTSV4RRFFQ69G5FAY"),
-};
-expectEqual("mismatched receiptId/claim", validateApprovalSubmitRequest(badReq), {
-  ok: false,
-  reason: "receiptId must match approvalToken.claim.receiptId",
-});
-
-const emptyKeyReq = { ...goodReq, idempotencyKey: "" };
-expectEqual("empty idempotencyKey rejected", validateApprovalSubmitRequest(emptyKeyReq), {
-  ok: false,
-  reason: "idempotencyKey must match /^[A-Za-z0-9_-]{1,128}$/",
-});
-
-const missingTokenIdReq = {
-  ...goodReq,
-  approvalToken: {
-    ...validToken,
-    tokenId: undefined,
-  },
-};
-expectEqual("missing tokenId rejected", validateApprovalSubmitRequest(missingTokenIdReq), {
-  ok: false,
-  reason: "approvalToken/tokenId: is required",
-});
-
-// ──────────────────────────────────────────────────────────────────────────
-header(12, "EventLsn safe-integer bound (writer ⟷ verifier round-trip)");
-// ──────────────────────────────────────────────────────────────────────────
-expectThrows(() => lsnFromV1Number(Number.MAX_SAFE_INTEGER + 1), /non-negative safe integer/);
-expectThrows(() => lsnFromV1Number(-1), /non-negative safe integer/);
-expectEqual(
-  "MAX_SAFE_INTEGER accepted",
-  lsnFromV1Number(Number.MAX_SAFE_INTEGER),
-  `v1:${Number.MAX_SAFE_INTEGER}`,
-);
-
-// ──────────────────────────────────────────────────────────────────────────
-header(13, "Bonus: golden eventHash literal (cross-language wire contract)");
-// ──────────────────────────────────────────────────────────────────────────
-const goldenRecord: AuditEventRecord = {
-  seqNo: lsnFromV1Number(0),
-  timestamp: new Date("2026-05-08T00:00:00.000Z"),
-  prevHash: GENESIS_PREV_HASH,
-  eventHash: GENESIS_PREV_HASH,
-  payload: { kind: "boot_marker", body: new TextEncoder().encode("boot") },
-};
-expectEqual(
-  "boot_marker eventHash matches golden vector",
-  computeAuditEventHash(goldenRecord),
-  "e27134d1b1641fb13747d9fac78aecc90d9d1385d04bfeea4a8a596fdb6101bb",
-);
-console.log(
-  `  ${ANSI.dim}A Go/Rust/Python verifier implementing the wire contract MUST produce the same hash.${ANSI.reset}`,
-);
-console.log(
-  `  ${ANSI.dim}Run testdata/verifier-reference.go to verify cross-language portability.${ANSI.reset}`,
-);
-
-// ──────────────────────────────────────────────────────────────────────────
-header(14, "FrozenArgs.fromCanonical rehydrates canonical JSON");
-// ──────────────────────────────────────────────────────────────────────────
-const rehydratedArgs = FrozenArgs.fromCanonical(argsA.canonicalJson);
-expectEqual(
-  "fromCanonical preserves canonical JSON",
-  rehydratedArgs.canonicalJson,
-  argsA.canonicalJson,
-);
-expectEqual("fromCanonical preserves hash", rehydratedArgs.hash, argsA.hash);
-expectThrows(
-  () => FrozenArgs.fromCanonical('{"b":2,"a":1}'),
-  /FrozenArgs\.fromCanonical: input is not canonical-form/,
-);
-
-// ──────────────────────────────────────────────────────────────────────────
-header(15, "ApiBootstrap codec: snake_case wire ↔ camelCase TS");
-// ──────────────────────────────────────────────────────────────────────────
-// Wire JSON is snake_case (v0 contract: `{ token, broker_url }`); the TS
-// runtime surface is camelCase, enforced by `style/useNamingConvention`.
-// The codec functions are the only place those two shapes meet.
-const bootstrapWire = { token: "tok-bootstrap-demo-abc", broker_url: "http://127.0.0.1:54321" };
-const bootstrapTs = apiBootstrapFromJson(bootstrapWire);
-expectEqual("decoded brokerUrl (camelCase)", bootstrapTs.brokerUrl, "http://127.0.0.1:54321");
-expectEqual("round-trip back to wire shape", apiBootstrapToJson(bootstrapTs), {
-  token: "tok-bootstrap-demo-abc",
-  broker_url: "http://127.0.0.1:54321",
-});
-expectThrows(
-  () =>
-    apiBootstrapFromJson({ token: "tok-bootstrap-demo-abc", brokerUrl: "http://127.0.0.1:54321" }),
-  /broker_url|brokerUrl/,
-);
-
-// ──────────────────────────────────────────────────────────────────────────
-header(16, "Loopback gate rejects DNS-rebinding probes");
-// ──────────────────────────────────────────────────────────────────────────
-expectEqual("Host Localhost:3000 allowed", isAllowedLoopbackHost("Localhost:3000"), true);
-expectEqual("Host localhost.evil.com rejected", isAllowedLoopbackHost("localhost.evil.com"), false);
-expectEqual(
-  "Remote IPv4-mapped loopback allowed",
-  isLoopbackRemoteAddress("::ffff:127.0.0.1"),
-  true,
-);
-expectEqual("Remote with port rejected", isLoopbackRemoteAddress("127.0.0.1:1234"), false);
-
-// ──────────────────────────────────────────────────────────────────────────
-header(17, "IPC wire codecs pin WebAuthn approval tokens and runtime guards");
-// ──────────────────────────────────────────────────────────────────────────
-const approvalSubmitWire = {
-  receiptId: validReceipt.id,
-  idempotencyKey: "submit-01",
-  approvalToken: signedApprovalTokenToJsonValue(validToken),
-};
-const decodedSubmit = approvalSubmitRequestFromJson(approvalSubmitWire);
-expectEqual(
-  "approvalSubmitRequestFromJson preserves caller-supplied notBefore",
-  decodedSubmit.approvalToken.notBefore,
-  validToken.notBefore,
-);
-expectEqual("decoded submit request validates", validateApprovalSubmitRequest(decodedSubmit), {
-  ok: true,
-});
-expectEqual(
-  "oversized assertion signature rejected before regex",
-  validateApprovalSubmitRequest({
-    ...goodReq,
-    approvalToken: {
-      ...validToken,
-      signature: {
-        ...validToken.signature,
-        signature: "A".repeat(MAX_WEBAUTHN_ASSERTION_FIELD_BYTES + 1),
-      },
-    },
-  }),
-  {
-    ok: false,
-    reason:
-      "approvalToken/signature/signature: approvalToken/signature/signature bytes exceeds budget: 16385 > 16384",
-  },
-);
-expectEqual(
-  "malformed WebAuthn assertion rejected",
-  validateApprovalSubmitRequest({
-    ...goodReq,
-    approvalToken: {
-      ...validToken,
-      signature: {
-        ...validToken.signature,
-        signature: "not base64!",
-      },
-    },
-  }),
-  {
-    ok: false,
-    reason:
-      "approvalToken/signature/signature: must be a canonical non-empty unpadded base64url string",
-  },
-);
-expectEqual(
-  "stream event guard accepts tuple values",
-  STREAM_EVENT_KIND_VALUES.every(isStreamEventKind),
-  true,
-);
-expectEqual(
-  "stream event guard rejects unknown value",
-  isStreamEventKind("receipt.deleted"),
-  false,
-);
-expectEqual("WS frame guard accepts tuple values", WS_FRAME_TYPE_VALUES.every(isWsFrameType), true);
-expectEqual("WS frame guard rejects unknown value", isWsFrameType("close"), false);
-
-const threadId = asThreadId("01ARZ3NDEKTSV4RRFFQ69G5FAY");
-const revision1 = asThreadSpecRevisionId("01BRZ3NDEKTSV4RRFFQ69G5FA0");
-const revision2 = asThreadSpecRevisionId("01BRZ3NDEKTSV4RRFFQ69G5FA1");
-const staleRevision = asThreadSpecRevisionId("01BRZ3NDEKTSV4RRFFQ69G5FA2");
-const threadSigner = asSignerIdentity("fran@example.com");
-const specContent = {
-  body: "Implement the thread protocol slice",
-  checklist: ["receipt v2", "audit events", "stream invalidation"],
-};
-const thread = {
-  id: threadId,
-  title: "Thread protocol",
-  status: "open" as const,
-  spec: {
-    revisionId: revision1,
-    threadId,
-    content: specContent,
-    contentHash: threadSpecContentHash(specContent),
-    authoredBy: threadSigner,
-    authoredAt: new Date("2026-05-08T18:00:00.000Z"),
-  },
-  externalRefs: {
-    sourceUrls: ["https://example.test/wuphf/743"],
-    entityIds: ["issue:743"],
-  },
-  taskIds: [validReceipt.taskId],
-  createdBy: threadSigner,
-  createdAt: new Date("2026-05-08T18:00:00.000Z"),
-  updatedAt: new Date("2026-05-08T18:01:00.000Z"),
-};
-
-// ──────────────────────────────────────────────────────────────────────────
-header(18, "Thread create + spec edit + status change round-trip through canonical JSON");
-// ──────────────────────────────────────────────────────────────────────────
-expectEqual("threadFromJson(threadToJson(thread))", threadFromJson(threadToJson(thread)), thread);
-expectEqual(
-  "thread_created audit body is canonical JSON bytes",
-  textDecoder
-    .decode(
-      threadAuditPayloadToBytes("thread_created", {
-        threadId,
-        title: "Thread protocol",
-        createdBy: threadSigner,
-        createdAt: thread.createdAt,
-        externalRefs: thread.externalRefs,
-      }),
-    )
-    .includes('"threadId"'),
-  true,
-);
-
-// ──────────────────────────────────────────────────────────────────────────
-header(19, "Thread spec OCC accepts the prior revision and rejects stale-base 409 shape");
-// ──────────────────────────────────────────────────────────────────────────
-const firstEdit: ThreadSpecEditedAuditPayload = {
-  threadId,
-  revisionId: revision1,
-  content: specContent,
-  contentHash: threadSpecContentHash(specContent),
-  authoredBy: threadSigner,
-  authoredAt: new Date("2026-05-08T18:00:00.000Z"),
-};
-const secondEdit: ThreadSpecEditedAuditPayload = {
-  ...firstEdit,
-  revisionId: revision2,
-  baseRevisionId: revision1,
-};
-const staleEdit: ThreadSpecEditedAuditPayload = {
-  ...firstEdit,
-  revisionId: revision2,
-  baseRevisionId: staleRevision,
-};
-expectEqual("OCC happy path", validateThreadSpecRevisionChain([firstEdit, secondEdit]), {
-  ok: true,
-});
-expectEqual(
-  "OCC stale-base rejected",
-  validateThreadSpecRevisionChain([firstEdit, staleEdit]).ok,
-  false,
-);
-
-// ──────────────────────────────────────────────────────────────────────────
-header(20, "Thread status fold and terminal-write-once invariant");
-// ──────────────────────────────────────────────────────────────────────────
-expectEqual(
-  "status fold happy path",
-  validateThreadStatusFold([
-    { kind: "thread_created", threadId },
-    {
-      kind: "thread_status_changed",
-      threadId,
-      fromStatus: "open",
-      toStatus: "in_progress",
-      changedBy: threadSigner,
-      changedAt: new Date("2026-05-08T18:02:00.000Z"),
-    },
-    {
-      kind: "thread_status_changed",
-      threadId,
-      fromStatus: "in_progress",
-      toStatus: "closed",
-      changedBy: threadSigner,
-      changedAt: new Date("2026-05-08T18:03:00.000Z"),
-    },
-  ]),
-  { ok: true },
-);
-expectEqual(
-  "terminal transition rejected",
-  validateThreadStatusFold([
-    { kind: "thread_created", threadId },
-    {
-      kind: "thread_status_changed",
-      threadId,
-      fromStatus: "open",
-      toStatus: "closed",
-      changedBy: threadSigner,
-      changedAt: new Date("2026-05-08T18:02:00.000Z"),
-    },
-    {
-      kind: "thread_status_changed",
-      threadId,
-      fromStatus: "closed",
-      toStatus: "in_progress",
-      changedBy: threadSigner,
-      changedAt: new Date("2026-05-08T18:03:00.000Z"),
-    },
-  ]).ok,
-  false,
-);
-
-// ──────────────────────────────────────────────────────────────────────────
-header(21, "ReceiptSnapshotV1 rejects threadId presence with path /threadId");
-// ──────────────────────────────────────────────────────────────────────────
-const v1WithThreadId = JSON.parse(receiptToJson(validReceipt)) as Record<string, unknown>;
-v1WithThreadId.threadId = threadId;
-expectThrows(() => receiptFromJson(JSON.stringify(v1WithThreadId)), /\/threadId/);
-
-// ──────────────────────────────────────────────────────────────────────────
-header(22, "ReceiptSnapshotV2 round-trips with and without threadId");
-// ──────────────────────────────────────────────────────────────────────────
-const v2WithThread: ReceiptSnapshot = { ...validReceipt, schemaVersion: 2, threadId };
-const v2WithoutThread: ReceiptSnapshot = { ...validReceipt, schemaVersion: 2 };
-expectEqual(
-  "v2 with threadId canonical bytes",
-  receiptToJson(receiptFromJson(receiptToJson(v2WithThread))),
-  receiptToJson(v2WithThread),
-);
-expectEqual(
-  "v2 without threadId canonical bytes",
-  receiptToJson(receiptFromJson(receiptToJson(v2WithoutThread))),
-  receiptToJson(v2WithoutThread),
-);
-
-// ──────────────────────────────────────────────────────────────────────────
-header(23, "thread_spec_edited validator catches forged contentHash");
-// ──────────────────────────────────────────────────────────────────────────
-expectEqual(
-  "forged contentHash rejected",
-  validateThreadSpecEditedAuditPayload({ ...firstEdit, contentHash: sha256Hex("forged") }).ok,
-  false,
-);
-
-// ──────────────────────────────────────────────────────────────────────────
-header(24, "Thread stream event tuple guards stay invalidation-only");
-// ──────────────────────────────────────────────────────────────────────────
-expectEqual("thread.created accepted", isStreamEventKind("thread.created"), true);
-expectEqual("thread.updated accepted", isStreamEventKind("thread.updated"), true);
-expectEqual(
-  "thread.pinned_approvals.changed accepted",
-  isStreamEventKind("thread.pinned_approvals.changed"),
-  true,
-);
-const threadStreamEvent: ThreadStreamEvent = {
-  id: "evt-thread-updated-demo",
-  kind: "thread.updated",
-  emittedAt: "2026-05-08T18:02:00.000Z",
-  payload: {
-    threadId,
-    headLsn: lsnFromV1Number(3),
-  },
-};
-expectEqual(
-  "thread stream event validator accepts invalidation",
-  validateThreadStreamEvent(threadStreamEvent),
-  {
-    ok: true,
-  },
-);
-expectEqual(
-  "thread stream event validator rejects payload data",
-  validateThreadStreamEvent({
-    ...threadStreamEvent,
-    payload: {
-      ...threadStreamEvent.payload,
-      content: "secret",
-    },
-  }).ok,
-  false,
-);
-
-header(25, "Cost event uses integer MicroUsd — float drift is a wire-shape break");
-// ──────────────────────────────────────────────────────────────────────────
-// The §15.A invariant sum(cost_events) == sum(by_agent) == sum(by_task) is only
-// decidable on integers. A 0.1+0.2 style accumulation across thousands of
-// events is exactly what breaks ledger reconciliation in production.
-const costEvent: CostEventAuditPayload = {
-  receiptId: asReceiptId("01ARZ3NDEKTSV4RRFFQ69G5FAV"),
-  agentSlug: asAgentSlug("primary"),
-  taskId: asTaskId("01BRZ3NDEKTSV4RRFFQ69G5FA0"),
-  providerKind: asProviderKind("anthropic"),
-  model: "claude-opus-4-7",
-  amountMicroUsd: asMicroUsd(2_500_000),
-  units: { inputTokens: 1_024, outputTokens: 512, cacheReadTokens: 0, cacheCreationTokens: 0 },
-  occurredAt: new Date("2026-05-08T18:03:00.000Z"),
-};
-expectEqual("typed cost_event validates", validateCostEventAuditPayload(costEvent), { ok: true });
-expectThrows(() => asMicroUsd(1.5), /non-negative safe integer/);
-expectThrows(() => asMicroUsd(-1), /non-negative safe integer/);
-// Brand bound = MAX_BUDGET_LIMIT_MICRO_USD ($1M); per-event cap
-// MAX_COST_EVENT_AMOUNT_MICRO_USD ($100) is enforced in the validator so a
-// rogue cost_event over $100 cannot dominate a daily budget even if the
-// MicroUsd brand accepted the value.
-const overEventCap = asMicroUsd(MAX_COST_EVENT_AMOUNT_MICRO_USD + 1);
-expectEqual(
-  "validator rejects amount over MAX_COST_EVENT_AMOUNT_MICRO_USD",
-  validateCostEventAuditPayload({ ...costEvent, amountMicroUsd: overEventCap }).ok,
-  false,
-);
-const costBytes = costAuditPayloadToBytes("cost_event", costEvent);
-const roundTrip = costAuditPayloadFromJsonValue(
-  "cost_event",
-  JSON.parse(textDecoder.decode(costBytes)),
-);
-expectEqual("cost_event round-trips through canonical bytes", roundTrip, costEvent);
-
-header(26, "Budget thresholds: ascending, deduplicated, bounded");
-// ──────────────────────────────────────────────────────────────────────────
-// Threshold arrays drive the reactor — a duplicate would re-fire the same
-// crossing; a descending sequence would skip thresholds during scan; an
-// unbounded array would let one budget pin the reactor on adversarial input.
-const goodBudget: BudgetSetAuditPayload = {
-  budgetId: asBudgetId("01ARZ3NDEKTSV4RRFFQ69G5FAZ"),
-  scope: "global",
-  limitMicroUsd: asMicroUsd(5_000_000),
-  thresholdsBps: [5_000, 8_000, 10_000],
-  setBy: asSignerIdentity("fran@example.com"),
-  setAt: new Date("2026-05-08T18:00:00.000Z"),
-};
-expectEqual("ascending unique thresholds accepted", validateBudgetSetAuditPayload(goodBudget), {
-  ok: true,
-});
-expectEqual(
-  "duplicate threshold rejected",
-  validateBudgetSetAuditPayload({ ...goodBudget, thresholdsBps: [5_000, 5_000] }).ok,
-  false,
-);
-expectEqual(
-  "descending threshold rejected",
-  validateBudgetSetAuditPayload({ ...goodBudget, thresholdsBps: [9_000, 5_000] }).ok,
-  false,
-);
-expectEqual(
-  "empty threshold list rejected",
-  validateBudgetSetAuditPayload({ ...goodBudget, thresholdsBps: [] }).ok,
-  false,
-);
-const overCapThresholds = Array.from(
-  { length: MAX_BUDGET_THRESHOLDS + 1 },
-  (_, i) => (i + 1) * 100,
-);
-expectEqual(
-  `more than ${MAX_BUDGET_THRESHOLDS} thresholds rejected`,
-  validateBudgetSetAuditPayload({ ...goodBudget, thresholdsBps: overCapThresholds }).ok,
-  false,
-);
-// global scope MUST have absent subjectId; agent/task MUST have matching brand.
-expectEqual(
-  "global scope with subjectId rejected",
-  validateBudgetSetAuditPayload({
-    ...goodBudget,
-    subjectId: "primary",
-  }).ok,
-  false,
-);
-expectEqual(
-  "agent scope without AgentSlug subjectId rejected",
-  validateBudgetSetAuditPayload({ ...goodBudget, scope: "agent" }).ok,
-  false,
-);
-// limit === 0 is the tombstone marker — must validate as ok (codec preserves 0).
-expectEqual(
-  "tombstone (limit=0) accepted as budget_set",
-  validateBudgetSetAuditPayload({ ...goodBudget, limitMicroUsd: asMicroUsd(0) }),
-  { ok: true },
-);
-
-header(27, "Threshold crossing payload carries budgetSetLsn so bumps re-arm");
-// ──────────────────────────────────────────────────────────────────────────
-// The reactor projection keys crossings by (budgetId, budgetSetLsn, thresholdBps).
-// Without budgetSetLsn, raising a budget would silently re-fire the existing
-// crossing rows; with it, a new budget_set LSN re-arms thresholds automatically.
-const crossing: BudgetThresholdCrossedAuditPayload = {
-  budgetId: asBudgetId("01ARZ3NDEKTSV4RRFFQ69G5FAZ"),
-  budgetSetLsn: "v1:5" as EventLsn,
-  thresholdBps: 5_000,
-  observedMicroUsd: asMicroUsd(2_500_000),
-  limitMicroUsd: asMicroUsd(5_000_000),
-  crossedAtLsn: "v1:4" as EventLsn,
-  crossedAt: new Date("2026-05-08T18:03:00.001Z"),
-};
-expectEqual(
-  "crossing payload with valid LSNs validates",
-  validateBudgetThresholdCrossedAuditPayload(crossing),
-  { ok: true },
-);
-expectEqual(
-  "crossing payload rejects malformed budgetSetLsn",
-  validateBudgetThresholdCrossedAuditPayload({ ...crossing, budgetSetLsn: "v0:bogus" as EventLsn })
-    .ok,
-  false,
-);
-expectEqual(
-  "crossing payload rejects out-of-range thresholdBps",
-  validateBudgetThresholdCrossedAuditPayload({ ...crossing, thresholdBps: 20_000 }).ok,
-  false,
-);
-// Canonical bytes for the crossing must reproduce the cost_event/budget_set chain
-// position: this is the same byte projection hashed by the audit chain.
-const crossingBytes = costAuditPayloadToBytes("budget_threshold_crossed", crossing);
-expectEqual(
-  "crossing canonical bytes parseable as canonical JSON",
-  canonicalJSON(JSON.parse(textDecoder.decode(crossingBytes))),
-  textDecoder.decode(crossingBytes),
-);
-
-header(28, "Cost ledger surface guards: brand guards, kind dispatch, closed enums");
-// ──────────────────────────────────────────────────────────────────────────
-// Public exports must each be exercised — these guards and dispatchers are
-// the public boundary; if any rot, downstream packages silently break.
-expectEqual("isMicroUsd accepts a valid MicroUsd", isMicroUsd(asMicroUsd(1_000)), true);
-expectEqual("isMicroUsd rejects 1.5", isMicroUsd(1.5), false);
-expectEqual("isMicroUsd rejects -1", isMicroUsd(-1), false);
-expectEqual("isBudgetId accepts ULID-shaped id", isBudgetId("01ARZ3NDEKTSV4RRFFQ69G5FAZ"), true);
-expectEqual("isBudgetId rejects lowercase", isBudgetId("01arz3ndektsv4rrffq69g5faz"), false);
-expectEqual("isBudgetScope accepts global|agent|task", isBudgetScope("agent"), true);
-expectEqual("isBudgetScope rejects bogus", isBudgetScope("user"), false);
-expectEqual("BUDGET_SCOPE_VALUES is the closed scope tuple", Array.from(BUDGET_SCOPE_VALUES), [
-  "global",
-  "agent",
-  "task",
-]);
-expectEqual(
-  "ProviderKind compatibility floor is public",
-  MINIMUM_PROTOCOL_VERSION_FOR_PROVIDER_KIND,
-  "cost-provider-kind-v1",
-);
-expectEqual("isCostAuditEventKind accepts cost_event", isCostAuditEventKind("cost_event"), true);
-expectEqual("isCostAuditEventKind rejects unknown", isCostAuditEventKind("budget_unset"), false);
-expectEqual(
-  "validateCostAuditPayloadForKind dispatches by kind",
-  validateCostAuditPayloadForKind("cost_event", costEvent),
-  { ok: true },
-);
-expectEqual(
-  "MAX_BUDGET_THRESHOLD_BPS bounds threshold space at 10000 (= 100%)",
-  MAX_BUDGET_THRESHOLD_BPS,
-  10_000,
-);
-expectEqual(
-  "MAX_COST_MODEL_BYTES bounds cost_event.model at 128 UTF-8 bytes",
-  MAX_COST_MODEL_BYTES,
-  128,
-);
-// costAuditPayloadToJsonValue is the plain-JSON projection used before
-// canonicalJSON encodes for the wire. Object identity-shape is locked here.
-const costJson = costAuditPayloadToJsonValue("cost_event", costEvent);
-expectEqual(
-  "costAuditPayloadToJsonValue projects amount as plain number",
-  typeof (costJson as Record<string, unknown>).amountMicroUsd,
-  "number",
-);
-expectEqual(
-  "costAuditPayloadToJsonValue projects occurredAt as ISO string",
-  (costJson as Record<string, unknown>).occurredAt,
-  "2026-05-08T18:03:00.000Z",
-);
-
-// ──────────────────────────────────────────────────────────────────────────
-header(29, "CredentialHandle is opaque and redacted");
-// ──────────────────────────────────────────────────────────────────────────
-const credentialHandle = createCredentialHandle({
-  id: asCredentialHandleId("cred_0123456789ABCDEFGHIJKLMNOPQRSTUV"),
-  agentId: asAgentId("agent_alpha"),
-  scope: asCredentialScope("openai"),
-});
-const credentialFixtureSecret = "fixture-secret-value-do-not-use-0000";
-const credentialJson = JSON.stringify(credentialHandle);
-expectEqual(
-  "CredentialHandle JSON carries only versioned id",
-  credentialJson,
-  '{"version":1,"id":"cred_0123456789ABCDEFGHIJKLMNOPQRSTUV"}',
-);
-expectEqual(
-  "credentialHandleToJson returns the wire shape",
-  credentialHandleToJson(credentialHandle),
-  {
-    version: 1,
-    id: asCredentialHandleId("cred_0123456789ABCDEFGHIJKLMNOPQRSTUV"),
-  },
-);
-expectEqual(
-  "CredentialHandle JSON omits secret",
-  credentialJson.includes(credentialFixtureSecret),
-  false,
-);
-expectEqual(
-  "CredentialHandle toString is redacted",
-  String(credentialHandle),
-  "CredentialHandle(<redacted>)",
-);
-expectEqual(
-  "CredentialHandle inspect is redacted",
-  inspect(credentialHandle),
-  "CredentialHandle { id: <redacted> }",
-);
-expectEqual(
-  "structuredClone(CredentialHandle) loses handle capability",
-  structuredClone(credentialHandle),
-  {},
-);
-expectThrows(
-  () =>
-    credentialHandleFromJson(JSON.parse(credentialJson), {
-      broker: {} as never,
-      agentId: asAgentId("agent_alpha"),
-      scope: asCredentialScope("openai"),
-    }),
-  /BrokerIdentity is required/,
-);
-
-header(30, "RunnerSpawnRequest and RunnerEvent reject drift at the broker boundary");
-const runnerSpawnJson = {
-  schemaVersion: 1,
-  kind: "claude-cli",
-  agentId: "agent_alpha",
-  credential: { version: 1, id: "cred_runner0123456789ABCDEFGHIJKLMN" },
-  providerRoute: {
-    credentialScope: "anthropic",
-    providerKind: "anthropic",
-  },
-  options: {
-    kind: "claude-cli",
-    extraArgs: ["--max-turns", "3"],
-  },
-  prompt: "Summarize the task state.",
-  model: "claude-sonnet-4-7",
-  taskId: "01ARZ3NDEKTSV4RRFFQ69G5FAW",
-  costCeilingMicroUsd: 2_500_000,
-};
-const runnerSpawn = runnerSpawnRequestFromJson(runnerSpawnJson);
-expectEqual(
-  "runner spawn request round-trips",
-  runnerSpawnRequestToJsonValue(runnerSpawn),
-  runnerSpawnJson,
-);
-expectThrows(
-  () => runnerSpawnRequestFromJson({ ...runnerSpawnJson, extra: true }),
-  /runnerSpawnRequest\/extra: is not allowed/,
-);
-expectThrows(
-  () => runnerSpawnRequestFromJson({ ...runnerSpawnJson, schemaVersion: 999 }),
-  /unsupported schemaVersion/,
-);
-const openAICompatSpawnJson = {
-  schemaVersion: 1,
-  kind: "openai-compat",
-  agentId: "agent_alpha",
-  credential: { version: 1, id: "cred_runner0123456789ABCDEFGHIJKLMN" },
-  options: {
-    kind: "openai-compat",
-    endpoint: "https://api.openai.com/v1/chat/completions",
-    headers: { "OpenAI-Beta": "assistants=v2" },
-    timeoutMs: 60_000,
-  },
-  prompt: "Summarize the task state.",
-};
-expectEqual(
-  "openai-compatible runner options round-trip through the wire shape",
-  runnerSpawnRequestToJsonValue(runnerSpawnRequestFromJson(openAICompatSpawnJson)),
-  openAICompatSpawnJson,
-);
-expectThrows(
-  () =>
-    runnerSpawnRequestFromJson({
-      ...openAICompatSpawnJson,
-      options: { kind: "openai-compat" },
-    }),
-  /runnerSpawnRequest.options.endpoint: is required/,
-);
-const runnerCostEvent = runnerEventFromJson({
-  schemaVersion: 1,
-  kind: "cost",
-  runnerId: "run_0123456789ABCDEFGHIJKLMNOPQRSTUV",
-  entry: {
-    agentSlug: "agent_alpha",
-    providerKind: "anthropic",
-    model: "claude-sonnet-4-7",
-    amountMicroUsd: 2048,
-    units: {
-      inputTokens: 1536,
-      outputTokens: 512,
-      cacheReadTokens: 0,
-      cacheCreationTokens: 0,
-    },
-    occurredAt: "2026-05-08T18:00:02.000Z",
-  },
-  at: "2026-05-08T18:00:02.000Z",
-});
-expectEqual(
-  "runner cost event emits canonical JSON values",
-  runnerEventToJsonValue(runnerCostEvent),
-  {
-    schemaVersion: 1,
-    kind: "cost",
-    runnerId: "run_0123456789ABCDEFGHIJKLMNOPQRSTUV",
-    entry: {
-      agentSlug: "agent_alpha",
-      providerKind: "anthropic",
-      model: "claude-sonnet-4-7",
-      amountMicroUsd: 2048,
-      units: {
-        inputTokens: 1536,
-        outputTokens: 512,
-        cacheReadTokens: 0,
-        cacheCreationTokens: 0,
-      },
-      occurredAt: "2026-05-08T18:00:02.000Z",
-    },
-    at: "2026-05-08T18:00:02.000Z",
-  },
-);
-expectThrows(
-  () =>
-    runnerEventFromJson({
-      schemaVersion: 1,
-      kind: "stdout",
-      runnerId: "run_0123456789ABCDEFGHIJKLMNOPQRSTUV",
-      chunk: "ok",
-      at: "2026-05-08T18:00:02Z",
-    }),
-  /ISO8601 UTC millisecond/,
-);
-expectEqual(
-  "runner failure code registry exposes stable machine codes",
-  RUNNER_FAILURE_CODE_VALUES.includes("runner_input_buffer_overflow"),
-  true,
-);
-expectEqual(
-  "runner failure code guard accepts stable machine codes",
-  isRunnerFailureCode("runner_input_buffer_overflow"),
-  true,
-);
-expectEqual(
-  "runner failure code guard rejects unknown values",
-  isRunnerFailureCode("bogus"),
-  false,
-);
-const runnerFailedEvent = runnerEventFromJson({
-  schemaVersion: 1,
-  kind: "failed",
-  runnerId: "run_0123456789ABCDEFGHIJKLMNOPQRSTUV",
-  error: "usage exceeded configured cost ceiling",
-  code: "cost_ceiling_exceeded",
-  at: "2026-05-08T18:00:03.000Z",
-});
-expectEqual("runner failed event carries stable code", runnerEventToJsonValue(runnerFailedEvent), {
-  schemaVersion: 1,
-  kind: "failed",
-  runnerId: "run_0123456789ABCDEFGHIJKLMNOPQRSTUV",
-  error: "usage exceeded configured cost ceiling",
-  code: "cost_ceiling_exceeded",
-  at: "2026-05-08T18:00:03.000Z",
-});
-expectThrows(
-  () =>
-    runnerEventFromJson({
-      schemaVersion: 1,
-      kind: "failed",
-      runnerId: "run_0123456789ABCDEFGHIJKLMNOPQRSTUV",
-      error: "provider returned a bad response",
-      code: "not_a_stable_code",
-      at: "2026-05-08T18:00:03.000Z",
-    }),
-  /RunnerFailureCode/,
-);
-
-header(31, "ApprovalRole helpers keep the closed enum importable at runtime");
-expectEqual("asApprovalRole accepts a known role", asApprovalRole("approver"), "approver");
-expectEqual("isApprovalRole rejects unknown roles", isApprovalRole("custom"), false);
-expectThrows(() => asApprovalRole("custom"), /ApprovalRole/);
-
-header(32, "ApprovalRequest artifact folds pending and decided approval state");
-const approvalReceipt = buildValidReceipt();
-const approvalEvidence = nonNull(approvalReceipt.approvals[0], "approvalReceipt.approvals[0]");
-const approvalRequest: ApprovalRequest = {
-  id: asApprovalRequestId("01HRQ7KZ7D4E6F8G9H0J1K2M3N"),
-  claim: approvalEvidence.signedToken.claim,
-  scope: approvalEvidence.signedToken.scope,
-  riskClass: "high",
-  threadId: asThreadId("01ARZ3NDEKTSV4RRFFQ69G5FAY"),
-  taskId: approvalReceipt.taskId,
-  receiptId: approvalReceipt.id,
-  requestedBy: asSignerIdentity("fran@example.com"),
-  requestedAt: new Date("2026-05-08T18:00:00.000Z"),
-  status: "approved",
-  decision: {
-    decision: "approve",
-    decidedBy: asSignerIdentity("approver@example.com"),
-    decidedAt: new Date("2026-05-08T18:05:00.000Z"),
-    token: approvalEvidence.signedToken,
-  },
-  schemaVersion: 1,
-};
-const approvalRequestWire = approvalRequestToJsonValue(approvalRequest);
-expectEqual(
-  "approval request round-trips through canonical JSON",
-  canonicalJSON(approvalRequestToJsonValue(approvalRequestFromJsonValue(approvalRequestWire))),
-  canonicalJSON(approvalRequestWire),
-);
-expectEqual(
-  "approval request JSON stays canonical",
-  approvalRequestToJson(approvalRequest),
-  canonicalJSON(approvalRequestWire),
-);
-expectThrows(
-  () => approvalRequestFromJsonValue({ ...approvalRequestWire, shadow: true }),
-  /shadow.*not allowed/,
-);
-expectThrows(
-  () =>
-    approvalRequestFromJsonValue({
-      ...approvalRequestWire,
-      request_id: "A".repeat(27),
-    }),
-  /ApprovalRequestId bytes/,
-);
-const approvedWithoutDecision = { ...approvalRequestWire };
-Reflect.deleteProperty(approvedWithoutDecision, "decision");
-expectThrows(() => approvalRequestFromJsonValue(approvedWithoutDecision), /decision.*required/);
-expectThrows(
-  () => approvalRequestFromJsonValue({ ...approvalRequestWire, status: "pending" }),
-  /decision.*absent/,
-);
-expectThrows(
-  () =>
-    approvalRequestFromJsonValue({
-      ...approvalRequestWire,
-      decision: {
-        decision: "approve",
-        decided_by: "approver@example.com",
-        decided_at: "2026-05-08T18:05:00.000Z",
-      },
-    }),
-  /token.*required/,
-);
-const tokenBoundToDifferentClaim = JSON.parse(canonicalJSON(approvalRequestWire)) as {
-  decision: {
-    token: {
-      claim: Record<string, unknown>;
-      scope: Record<string, unknown>;
-    };
-  };
-};
-tokenBoundToDifferentClaim.decision.token.claim.claimId = "claim_demo_receipt_cosign_02";
-tokenBoundToDifferentClaim.decision.token.scope.claimId = "claim_demo_receipt_cosign_02";
-expectThrows(
-  () => approvalRequestFromJsonValue(tokenBoundToDifferentClaim),
-  /token.*claim.*must match request claim/,
-);
-expectThrows(
-  () =>
-    approvalRequestFromJsonValue({
-      ...approvalRequestWire,
-      receipt_id: "01ARZ3NDEKTSV4RRFFQ69G5FAZ",
-    }),
-  /receiptId.*must match claim\.receiptId/,
-);
-
-header(33, "Route-envelope codecs own thread and approval HTTP bodies");
-const routeThreadId = asThreadId("01ARZ3NDEKTSV4RRFFQ69G5FAY");
-const routeRevision1 = asThreadSpecRevisionId("01BRZ3NDEKTSV4RRFFQ69G5FA0");
-const routeRevision2 = asThreadSpecRevisionId("01BRZ3NDEKTSV4RRFFQ69G5FA1");
-const routeCreatedAt = new Date("2026-05-08T18:00:00.000Z");
-const routeUpdatedAt = new Date("2026-05-08T18:05:00.000Z");
-const routeSpecContent = {
-  body: "Implement route envelope codecs",
-  checklist: ["codecs", "vectors", "demo"],
-};
-const routeEditedContent = { body: "Edited", checklist: ["tests", "vectors"] };
-const routeThread: Thread = {
-  id: routeThreadId,
-  title: "Approval request protocol",
-  status: "open",
-  spec: {
-    revisionId: routeRevision1,
-    threadId: routeThreadId,
-    content: routeSpecContent,
-    contentHash: threadSpecContentHash(routeSpecContent),
-    authoredBy: asSignerIdentity("fran@example.com"),
-    authoredAt: routeCreatedAt,
-  },
-  externalRefs: {
-    sourceUrls: ["https://example.test/wuphf/914"],
-    entityIds: ["issue:914"],
-  },
-  taskIds: [approvalReceipt.taskId],
-  createdBy: asSignerIdentity("fran@example.com"),
-  createdAt: routeCreatedAt,
-  updatedAt: routeUpdatedAt,
-};
-const routeThreadJson = JSON.parse(threadToJson(routeThread));
-const pendingApprovalRequestWire = { ...approvalRequestWire, status: "pending" };
-Reflect.deleteProperty(pendingApprovalRequestWire, "decision");
-const pendingApprovalRequest = approvalRequestFromJsonValue(pendingApprovalRequestWire);
-
-expectCodecRoundTrip(
-  "thread create request route envelope",
-  {
-    title: routeThread.title,
-    specContent: routeSpecContent,
-    externalRefs: { source_urls: routeThread.externalRefs.sourceUrls, entity_ids: [] },
-    idempotencyKey: asIdempotencyKey("route-envelope-demo-01"),
-  },
-  threadCreateRequestFromJson,
-  threadCreateRequestToJsonValue,
-);
-expectCodecRoundTrip(
-  "thread spec edit request route envelope",
-  {
-    schemaVersion: 1,
-    baseRevisionId: routeRevision1,
-    baseContentHash: routeThread.spec.contentHash,
-    content: routeEditedContent,
-    idempotencyKey: asIdempotencyKey("route-envelope-demo-02"),
-  },
-  threadSpecEditRequestFromJson,
-  threadSpecEditRequestToJsonValue,
-);
-expectCodecRoundTrip(
-  "thread status change request route envelope",
-  {
-    schemaVersion: 1,
-    fromStatus: "open",
-    toStatus: "in_progress",
-    idempotencyKey: asIdempotencyKey("route-envelope-demo-03"),
-  },
-  threadStatusChangeRequestFromJson,
-  threadStatusChangeRequestToJsonValue,
-);
-expectCodecRoundTrip(
-  "thread mutation response route envelope",
-  {
-    schemaVersion: 1,
-    threadId: routeThreadId,
-    headLsn: lsnFromV1Number(42),
-    revisionId: routeRevision2,
-    contentHash: threadSpecContentHash(routeEditedContent),
-  },
-  threadMutationResponseFromJson,
-  threadMutationResponseToJsonValue,
-);
-expectCodecRoundTrip(
-  "thread list response route envelope",
-  {
-    schemaVersion: 1,
-    threads: [routeThreadJson],
-    nextCursor: "bHNuOjQy",
-  },
-  threadListResponseFromJson,
-  threadListResponseToJsonValue,
-);
-expectCodecRoundTrip(
-  "thread get response route envelope",
-  {
-    schemaVersion: 1,
-    thread: routeThreadJson,
-  },
-  threadGetResponseFromJson,
-  threadGetResponseToJsonValue,
-);
-expectCodecRoundTrip(
-  "approval request create route envelope",
-  {
-    schemaVersion: 1,
-    claim: approvalEvidence.signedToken.claim,
-    scope: approvalEvidence.signedToken.scope,
-    riskClass: "high",
-    threadId: routeThreadId,
-    taskId: approvalReceipt.taskId,
-    receiptId: approvalReceipt.id,
-    idempotencyKey: asIdempotencyKey("route-envelope-demo-04"),
-  },
-  approvalRequestCreateRequestFromJson,
-  approvalRequestCreateRequestToJsonValue,
-);
-expectCodecRoundTrip(
-  "approval decision request route envelope",
-  {
-    schemaVersion: 1,
-    decision: "approve",
-    token: signedApprovalTokenToJsonValue(approvalEvidence.signedToken),
-    idempotencyKey: asIdempotencyKey("route-envelope-demo-05"),
-  },
-  approvalDecisionRequestFromJson,
-  approvalDecisionRequestToJsonValue,
-);
-expectCodecRoundTrip(
-  "approval request create response route envelope",
-  {
-    schemaVersion: 1,
-    approvalRequest: approvalRequestToJsonValue(pendingApprovalRequest),
-    headLsn: lsnFromV1Number(42),
-  },
-  approvalRequestCreateResponseFromJson,
-  approvalRequestCreateResponseToJsonValue,
-);
-expectCodecRoundTrip(
-  "approval decision response route envelope",
-  {
-    schemaVersion: 1,
-    approvalRequest: approvalRequestWire,
-    headLsn: lsnFromV1Number(43),
-  },
-  approvalDecisionResponseFromJson,
-  approvalDecisionResponseToJsonValue,
-);
-expectCodecRoundTrip(
-  "route error envelope",
-  {
-    error: "store_busy",
-    message: "The projection store is temporarily busy.",
-    retryAfterMs: 1000,
-  },
-  routeErrorFromJson,
-  routeErrorToJsonValue,
-);
-
-// ──────────────────────────────────────────────────────────────────────────
-console.log("");
-console.log(`${ANSI.bold}─────────────────────────────────────${ANSI.reset}`);
-const summaryColor = failed === 0 ? ANSI.green : ANSI.red;
-console.log(`${ANSI.bold}${summaryColor}${passed} passed, ${failed} failed${ANSI.reset}`);
-process.exit(failed === 0 ? 0 : 1);
-
-// ──────────────────────────────────────────────────────────────────────────
-function buildValidReceipt(): ReceiptSnapshot {
-  const receiptId = asReceiptId("01ARZ3NDEKTSV4RRFFQ69G5FAV");
-  const writeId = asWriteId("write_01");
-  const proposedDiff = FrozenArgs.freeze({ amount: { from: 1000, to: 1500 } });
-  const claimId = asApprovalClaimId("claim_demo_receipt_cosign_01");
-  const signedToken = {
-    schemaVersion: 1 as const,
-    tokenId: asApprovalTokenId("01HX6P2D8T4Y7K9M3N5Q1R6S2V"),
-    claim: {
-      schemaVersion: 1 as const,
-      claimId,
-      kind: "receipt_co_sign" as const,
-      receiptId,
-      writeId,
-      frozenArgsHash: proposedDiff.hash,
-      riskClass: "high" as const,
-    },
-    scope: {
-      mode: "single_use" as const,
-      claimId,
-      claimKind: "receipt_co_sign" as const,
-      role: "approver" as const,
-      maxUses: 1 as const,
-      receiptId,
-      writeId,
-      frozenArgsHash: proposedDiff.hash,
-    },
-    notBefore: asTimestampMs(Date.UTC(2026, 4, 8, 18, 0, 0, 0)),
-    expiresAt: asTimestampMs(Date.UTC(2026, 4, 8, 18, 30, 0, 0)),
-    issuedTo: asAgentId("agent_alpha"),
-    signature: {
-      credentialId: "Y3JlZGVudGlhbC0wMQ",
-      authenticatorData: "YXV0aGVudGljYXRvci1kYXRh",
-      clientDataJson: "Y2xpZW50LWRhdGEtanNvbg",
-      signature: "ZmFrZS13ZWJhdXRobi1zaWduYXR1cmU",
-      userHandle: "dXNlci0wMQ",
-    },
-  };
-  return {
-    id: receiptId,
-    agentSlug: asAgentSlug("sam_agent"),
-    taskId: asTaskId("01ARZ3NDEKTSV4RRFFQ69G5FAW"),
-    triggerKind: "human_message",
-    triggerRef: "msg:01ARZ3NDEKTSV4RRFFQ69G5FAX",
-    startedAt: new Date("2026-05-08T18:00:00.000Z"),
-    finishedAt: new Date("2026-05-08T18:05:00.000Z"),
-    status: "ok",
-    providerKind: asProviderKind("openai"),
-    model: "gpt-5.2",
-    promptHash: sha256Hex("prompt:v1"),
-    toolManifest: sha256Hex("tool-manifest:v1"),
-    toolCalls: [
-      {
-        toolId: asToolCallId("tool_01"),
-        toolName: "hubspot.deals.update",
-        inputs: FrozenArgs.freeze({ deal: "5678", action: "advance_stage" }),
-        output: SanitizedString.fromUnknown("Stage advanced to qualified"),
-        startedAt: new Date("2026-05-08T18:00:01.000Z"),
-        finishedAt: new Date("2026-05-08T18:00:02.000Z"),
-        status: "ok",
-        error: SanitizedString.fromUnknown(""),
-      },
-    ],
-    approvals: [
-      {
-        approvalId: asApprovalId("approval_01"),
-        role: "approver",
-        decision: "approve",
-        signedToken,
-        tokenVerdict: { status: "valid", verifiedAt: new Date("2026-05-08T18:01:00.000Z") },
-        decidedAt: new Date("2026-05-08T18:01:00.000Z"),
-      },
-    ],
-    filesChanged: [],
-    commits: [],
-    sourceReads: [],
-    writes: [
-      {
-        writeId,
-        action: "hubspot.deals.update",
-        target: "deal:5678",
-        idempotencyKey: asIdempotencyKey("write-01"),
-        proposedDiff,
-        appliedDiff: FrozenArgs.freeze({ stage: { from: "lead", to: "qualified" } }),
-        approvalToken: signedToken,
-        approvedAt: new Date("2026-05-08T18:01:01.000Z"),
-        result: "applied",
-        postWriteVerify: FrozenArgs.freeze({ stage: "qualified" }),
-      },
-    ],
-    inputTokens: 1200,
-    outputTokens: 345,
-    cacheReadTokens: 50,
-    cacheCreationTokens: 25,
-    costUsd: 0.0425,
-    finalMessage: SanitizedString.fromUnknown("Done."),
-    error: SanitizedString.fromUnknown(""),
-    notebookWrites: [],
-    wikiWrites: [],
-    schemaVersion: 1,
-  };
-}
-
-// Defeat unused-import warnings for symbols the demo references via dynamic
-// scenarios above (kept here so the import block stays a complete map of
-// what the package exposes for adversarial play).
-void canonicalJSON;
-void asApiToken;
-void asMerkleRootHex;
-void serializeAuditEventRecordForHash;
-void validateMerkleRootRecord;
+printSummaryAndExit();
