+++ packages/protocol/scripts/demo/fixtures.ts
@@ -1,0 +1,121 @@
+import {
+  asAgentId,
+  asAgentSlug,
+  asApprovalClaimId,
+  asApprovalId,
+  asApprovalTokenId,
+  asIdempotencyKey,
+  asProviderKind,
+  asReceiptId,
+  asTaskId,
+  asTimestampMs,
+  asToolCallId,
+  asWriteId,
+  FrozenArgs,
+  type ReceiptSnapshot,
+  SanitizedString,
+  sha256Hex,
+} from "../../src/index.ts";
+
+export function buildValidReceipt(): ReceiptSnapshot {
+  const receiptId = asReceiptId("01ARZ3NDEKTSV4RRFFQ69G5FAV");
+  const writeId = asWriteId("write_01");
+  const proposedDiff = FrozenArgs.freeze({ amount: { from: 1000, to: 1500 } });
+  const claimId = asApprovalClaimId("claim_demo_receipt_cosign_01");
+  const signedToken = {
+    schemaVersion: 1 as const,
+    tokenId: asApprovalTokenId("01HX6P2D8T4Y7K9M3N5Q1R6S2V"),
+    claim: {
+      schemaVersion: 1 as const,
+      claimId,
+      kind: "receipt_co_sign" as const,
+      receiptId,
+      writeId,
+      frozenArgsHash: proposedDiff.hash,
+      riskClass: "high" as const,
+    },
+    scope: {
+      mode: "single_use" as const,
+      claimId,
+      claimKind: "receipt_co_sign" as const,
+      role: "approver" as const,
+      maxUses: 1 as const,
+      receiptId,
+      writeId,
+      frozenArgsHash: proposedDiff.hash,
+    },
+    notBefore: asTimestampMs(Date.UTC(2026, 4, 8, 18, 0, 0, 0)),
+    expiresAt: asTimestampMs(Date.UTC(2026, 4, 8, 18, 30, 0, 0)),
+    issuedTo: asAgentId("agent_alpha"),
+    signature: {
+      credentialId: "Y3JlZGVudGlhbC0wMQ",
+      authenticatorData: "YXV0aGVudGljYXRvci1kYXRh",
+      clientDataJson: "Y2xpZW50LWRhdGEtanNvbg",
+      signature: "ZmFrZS13ZWJhdXRobi1zaWduYXR1cmU",
+      userHandle: "dXNlci0wMQ",
+    },
+  };
+  return {
+    id: receiptId,
+    agentSlug: asAgentSlug("sam_agent"),
+    taskId: asTaskId("01ARZ3NDEKTSV4RRFFQ69G5FAW"),
+    triggerKind: "human_message",
+    triggerRef: "msg:01ARZ3NDEKTSV4RRFFQ69G5FAX",
+    startedAt: new Date("2026-05-08T18:00:00.000Z"),
+    finishedAt: new Date("2026-05-08T18:05:00.000Z"),
+    status: "ok",
+    providerKind: asProviderKind("openai"),
+    model: "gpt-5.2",
+    promptHash: sha256Hex("prompt:v1"),
+    toolManifest: sha256Hex("tool-manifest:v1"),
+    toolCalls: [
+      {
+        toolId: asToolCallId("tool_01"),
+        toolName: "hubspot.deals.update",
+        inputs: FrozenArgs.freeze({ deal: "5678", action: "advance_stage" }),
+        output: SanitizedString.fromUnknown("Stage advanced to qualified"),
+        startedAt: new Date("2026-05-08T18:00:01.000Z"),
+        finishedAt: new Date("2026-05-08T18:00:02.000Z"),
+        status: "ok",
+        error: SanitizedString.fromUnknown(""),
+      },
+    ],
+    approvals: [
+      {
+        approvalId: asApprovalId("approval_01"),
+        role: "approver",
+        decision: "approve",
+        signedToken,
+        tokenVerdict: { status: "valid", verifiedAt: new Date("2026-05-08T18:01:00.000Z") },
+        decidedAt: new Date("2026-05-08T18:01:00.000Z"),
+      },
+    ],
+    filesChanged: [],
+    commits: [],
+    sourceReads: [],
+    writes: [
+      {
+        writeId,
+        action: "hubspot.deals.update",
+        target: "deal:5678",
+        idempotencyKey: asIdempotencyKey("write-01"),
+        proposedDiff,
+        appliedDiff: FrozenArgs.freeze({ stage: { from: "lead", to: "qualified" } }),
+        approvalToken: signedToken,
+        approvedAt: new Date("2026-05-08T18:01:01.000Z"),
+        result: "applied",
+        postWriteVerify: FrozenArgs.freeze({ stage: "qualified" }),
+      },
+    ],
+    inputTokens: 1200,
+    outputTokens: 345,
+    cacheReadTokens: 50,
+    cacheCreationTokens: 25,
+    costUsd: 0.0425,
+    finalMessage: SanitizedString.fromUnknown("Done."),
+    error: SanitizedString.fromUnknown(""),
+    notebookWrites: [],
+    wikiWrites: [],
+    schemaVersion: 1,
+  };
+}
