+++ packages/protocol/scripts/demo/public-surface-sentinels.ts
@@ -1,0 +1,13 @@
+import {
+  asApiToken,
+  asMerkleRootHex,
+  serializeAuditEventRecordForHash,
+  validateMerkleRootRecord,
+} from "../../src/index.ts";
+
+export function touchPublicSurfaceSentinels(): void {
+  void asApiToken;
+  void asMerkleRootHex;
+  void serializeAuditEventRecordForHash;
+  void validateMerkleRootRecord;
+}
